Move fluent Firestore classes to lop level
