+++ gnd/src/main/java/com/google/android/gnd/persistence/remote/firestore/GndFirestore.java
@@ -29,6 +29,8 @@ import com.google.android.gnd.model.observation.Observation;
 import com.google.android.gnd.model.observation.ObservationMutation;
 import com.google.android.gnd.persistence.remote.RemoteDataEvent;
 import com.google.android.gnd.persistence.remote.firestore.fluent.AbstractFluentFirestore;
+import com.google.android.gnd.persistence.remote.firestore.fluent.FluentCollectionReference;
+import com.google.android.gnd.persistence.remote.firestore.fluent.FluentDocumentReference;
 import com.google.common.collect.ImmutableList;
 import com.google.firebase.firestore.CollectionReference;
 import com.google.firebase.firestore.DocumentChange;
